Add V2 Boards and Metadata Interface
Remove the horribly broken code which erases the factory-provided calibration of the
hardware.

Remove DATE, TIME macros. The build should be reproducible, creating the same binary
output, not depend on 'random' values.

Fix build errors with deprecated ASM instructions.

Add Versio Duo and ACPAD board definitions.

Export a JSON record like:
{
  "com.versioduo.bootloader": {
    "board": "versioduo:samd:express-mini"
  }
}

to be read from the running sketch, which will refuse to update itself
with the firmware for a different board.